fix(benchmark-test): suggestions for #10975
