+++ packages/benchmark-test/rollup.config.js
@@ -2,16 +2,12 @@ import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 
 export default {
-  input: 'benchmark.test.js', 
+  input: 'test/benchmark.test.js',
   output: {
     file: 'dist/bundle.js',
     format: 'iife',
     name: 'bundle',
     sourcemap: false,
   },
-  plugins: [
-    resolve(),
-    commonjs(), 
-  ],
+  plugins: [resolve(), commonjs()],
 };
-
