+++ packages/benchmark-test/tsconfig.build.json
@@ -1,0 +1,6 @@
+{
+    "extends": [
+        "./tsconfig.json",
+        "../../tsconfig-build-options.json"
+    ]
+}
