+++ packages/benchmark-test/tsconfig.json
@@ -1,0 +1,10 @@
+// This file can contain .js-specific Typescript compiler config.
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {},
+  "include": [
+    "tools",
+    "test",
+    "rollup.config.js"
+  ]
+}
